Moved functions from Rule to SourceCodeUnit. Moved functions from utilities to Rule
Replace getters/setters with getset macro
