+++ polyglot/piranha/src/lib.rs
@@ -30,26 +30,11 @@ use itertools::Itertools;
 use jwalk::WalkDir;
 use log::{debug, info};
 use regex::Regex;
-use tree_sitter::{Parser, Range};
+use tree_sitter::Parser;
 
-use crate::{
-  models::rule_store::RuleStore,
-  utilities::{read_file, tree_sitter_utilities::get_replace_range},
-};
-
-use crate::models::scopes::ScopeGenerator;
-use crate::utilities::tree_sitter_utilities::PiranhaHelpers;
-use crate::{
-  models::{
-    rule::Rule,
-    rule_store::{GLOBAL, PARENT},
-  },
-  utilities::tree_sitter_utilities::get_node_for_range,
-};
+use crate::{models::rule_store::RuleStore, utilities::read_file};
 
 use pyo3::prelude::{pyfunction, pymodule, wrap_pyfunction, PyModule, PyResult, Python};
-use std::collections::VecDeque;
-use tree_sitter::Node;
 
 /// Executes piranha for the provided configuration at {path_to_configurations} upon the given {path_to_codebase}.
 ///
@@ -80,9 +65,7 @@ fn polyglot_piranha(_py: Python<'_>, m: &PyModule) -> PyResult<()> {
   Ok(())
 }
 
-pub fn execute_piranha(
-  configuration: &PiranhaArguments,
-) -> Vec<PiranhaOutputSummary> {
+pub fn execute_piranha(configuration: &PiranhaArguments) -> Vec<PiranhaOutputSummary> {
   info!("Executing Polyglot Piranha !!!");
 
   let mut flag_cleaner = FlagCleaner::new(configuration);
@@ -107,245 +90,20 @@ pub fn execute_piranha(
 }
 
 fn log_piranha_output_summaries(summaries: &Vec<PiranhaOutputSummary>) {
-    let mut total_number_of_matches : usize = 0;
-    let mut total_number_of_rewrites : usize = 0;
-    for summary in summaries {
-      let number_of_rewrites =&summary.rewrites().len();
-      let number_of_matches = &summary.matches().len();
-      info!("File : {:?}", &summary.path());
-      info!("  # Rewrites : {}", number_of_rewrites);
-      info!("  # Matches : {}", number_of_matches);
-      total_number_of_rewrites += number_of_rewrites;
-      total_number_of_matches += number_of_matches;
-    }
-    info!("Total files affected/matched {}", &summaries.len());
-    info!("Total number of matches {}", total_number_of_matches);
-    info!("Total number of rewrites {}", total_number_of_rewrites);
-}
-
-impl SourceCodeUnit {
-  /// Will apply the `rule` to all of its occurrences in the source code unit.
-  fn apply_rule(
-    &mut self, rule: Rule, rules_store: &mut RuleStore, parser: &mut Parser,
-    scope_query: &Option<String>,
-  ) {
-    loop {
-      if !self._apply_rule(rule.clone(), rules_store, parser, scope_query) {
-        break;
-      }
-    }
+  let mut total_number_of_matches: usize = 0;
+  let mut total_number_of_rewrites: usize = 0;
+  for summary in summaries {
+    let number_of_rewrites = &summary.rewrites().len();
+    let number_of_matches = &summary.matches().len();
+    info!("File : {:?}", &summary.path());
+    info!("  # Rewrites : {}", number_of_rewrites);
+    info!("  # Matches : {}", number_of_matches);
+    total_number_of_rewrites += number_of_rewrites;
+    total_number_of_matches += number_of_matches;
   }
-
-  /// Applies the rule to the first match in the source code
-  /// This is implements the main algorithm of piranha.
-  /// Parameters:
-  /// * `rule` : the rule to be applied
-  /// * `rule_store`: contains the input rule graph.
-  ///
-  /// Algorithm:
-  /// * check if the rule is match only
-  /// ** IF not (i.e. it is a rewrite):
-  /// *** Get the first match of the rule for the file
-  ///  (We only get the first match because the idea is that we will apply this change, and keep calling this method `_apply_rule` until all
-  /// matches have been exhaustively updated.
-  /// *** Apply the rewrite
-  /// *** Update the substitution table
-  /// *** Propagate the change
-  /// ** Else (i.e. it is a match only rule):
-  /// *** Get all the matches, and for each match
-  /// *** Update the substitution table
-  /// *** Propagate the change
-  fn _apply_rule(
-    &mut self, rule: Rule, rule_store: &mut RuleStore, parser: &mut Parser,
-    scope_query: &Option<String>,
-  ) -> bool {
-    let scope_node = self.get_scope_node(scope_query, rule_store);
-
-    let mut query_again = false;
-
-    // When rule is a "rewrite" rule :
-    // Update the first match of the rewrite rule
-    // Add mappings to the substitution
-    // Propagate each applied edit. The next rule will be applied relative to the application of this edit.
-    if !rule.is_match_only_rule() {
-      if let Some(edit) = rule.get_edit(&self.clone(), rule_store, scope_node, true) {
-        self.rewrites_mut().push(edit.clone());
-        query_again = true;
-
-        // Add all the (code_snippet, tag) mapping to the substitution table.
-        self.add_to_substitutions(edit.matches(), rule_store);
-
-        // Apply edit_1
-        let applied_ts_edit = self.apply_edit(&edit, parser);
-
-        self.propagate(get_replace_range(applied_ts_edit), rule, rule_store, parser);
-      }
-    }
-    // When rule is a "match-only" rule :
-    // Get all the matches
-    // Add mappings to the substitution
-    // Propagate each match. Note that,  we pass a identity edit (where old range == new range) in to the propagate logic.
-    // The next edit will be applied relative to the identity edit.
-    else {
-      for m in rule.get_matches(&self.clone(), rule_store, scope_node, true) {
-        self.matches_mut().push((rule.name(), m.clone()));
-
-        // In this scenario we pass the match and replace range as the range of the match `m`
-        // This is equivalent to propagating an identity rule
-        //  i.e. a rule that replaces the matched code with itself
-        // Note that, here we DO NOT invoke the `_apply_edit` method and only update the `substitutions`
-        // By NOT invoking this we simulate the application of an identity rule
-        //
-        self.add_to_substitutions(m.matches(), rule_store);
-
-        self.propagate(m.range(), rule.clone(), rule_store, parser);
-      }
-    }
-    query_again
-  }
-
-  /// This is the propagation logic of the Piranha's main algorithm.
-  /// Parameters:
-  ///  * `applied_ts_edit` -  it's(`rule`'s) application site (in terms of replacement range)
-  ///  * `rule` - The `rule` that was just applied
-  ///  * `rule_store` - contains the input "rule graph"
-  ///  * `parser` - parser for the language
-  /// Algorithm:
-  ///
-  /// (i) Lookup the `rule_store` and get all the (next) rules that could be after applying the current rule (`rule`).
-  ///   * We will receive the rules grouped by scope:  `GLOBAL` and `PARENT` are applicable to each language. However, other scopes are determined
-  ///     based on the `<language>/scope_config.toml`.
-  /// (ii) Add the `GLOBAL` rule to the global rule list in the `rule_store` (This will be performed in the next iteration)
-  /// (iii) Apply the local cleanup i.e. `PARENT` scoped rules
-  ///  (iv) Go to step 1 (and repeat this for the applicable parent scoped rule. Do this until, no parent scoped rule is applicable.) (recursive)
-  ///  (iv) Apply the rules based on custom language specific scopes (as defined in `<language>/scope_config.toml`) (recursive)
-  ///
-  fn propagate(
-    &mut self, replace_range: Range, rule: Rule, rules_store: &mut RuleStore, parser: &mut Parser,
-  ) {
-    let mut current_replace_range = replace_range;
-
-    let mut current_rule = rule.name();
-    let mut next_rules_stack: VecDeque<(String, Rule)> = VecDeque::new();
-    // Perform the parent edits, while queueing the Method and Class level edits.
-    // let file_level_scope_names = [METHOD, CLASS];
-    loop {
-      // Get all the (next) rules that could be after applying the current rule (`rule`).
-      let next_rules_by_scope = rules_store.get_next(&current_rule, self.substitutions());
-
-      debug!(
-        "\n{}",
-        &next_rules_by_scope
-          .iter()
-          .map(|(k, v)| {
-            let rules = v.iter().map(|f| f.name()).join(", ");
-            format!("Next Rules:\nScope {k} \nRules {rules}").blue()
-          })
-          .join("\n")
-      );
-
-      // Adds rules of scope != ["Parent", "Global"] to the stack
-      self.add_rules_to_stack(
-        &next_rules_by_scope,
-        current_replace_range,
-        rules_store,
-        &mut next_rules_stack,
-      );
-
-      // Add Global rules as seed rules
-      for r in &next_rules_by_scope[GLOBAL] {
-        rules_store.add_to_global_rules(r, self.substitutions());
-      }
-
-      // Process the parent
-      // Find the rules to be applied in the "Parent" scope that match any parent (context) of the changed node in the previous edit
-      if let Some(edit) = Rule::get_edit_for_context(
-        &self.clone(),
-        current_replace_range.start_byte,
-        current_replace_range.end_byte,
-        rules_store,
-        &next_rules_by_scope[PARENT],
-      ) {
-        self.rewrites_mut().push(edit.clone());
-        debug!(
-          "\n{}",
-          format!(
-            "Cleaning up the context, by applying the rule - {}",
-            edit.matched_rule()
-          )
-          .green()
-        );
-        // Apply the matched rule to the parent
-        let applied_edit = self.apply_edit(&edit, parser);
-        current_replace_range = get_replace_range(applied_edit);
-        current_rule = edit.matched_rule();
-        // Add the (tag, code_snippet) mapping to substitution table.
-        self.add_to_substitutions(edit.matches(), rules_store);
-      } else {
-        // No more parents found for cleanup
-        break;
-      }
-    }
-
-    // Apply the next rules from the stack
-    for (sq, rle) in &next_rules_stack {
-      self.apply_rule(rle.clone(), rules_store, parser, &Some(sq.to_string()));
-    }
-  }
-
-  /// Adds the "Method" and "Class" scoped next rules to the queue.
-  fn add_rules_to_stack(
-    &mut self, next_rules_by_scope: &HashMap<String, Vec<Rule>>, current_match_range: Range,
-    rules_store: &mut RuleStore, stack: &mut VecDeque<(String, Rule)>,
-  ) {
-    for (scope_level, rules) in next_rules_by_scope {
-      // Scope level is not "PArent" or "Global"
-      if ![PARENT, GLOBAL].contains(&scope_level.as_str()) {
-        for rule in rules {
-          let scope_query = ScopeGenerator::get_scope_query(
-            self.clone(),
-            scope_level,
-            current_match_range.start_byte,
-            current_match_range.end_byte,
-            rules_store,
-          );
-          // Add Method and Class scoped rules to the queue
-          stack.push_front((scope_query, rule.instantiate(self.substitutions())));
-        }
-      }
-    }
-  }
-
-  fn get_scope_node(&self, scope_query: &Option<String>, rules_store: &mut RuleStore) -> Node {
-    // Get scope node
-    // let mut scope_node = self.root_node();
-    if let Some(query_str) = scope_query {
-      // Apply the scope query in the source code and get the appropriate node
-      let tree_sitter_scope_query = rules_store.query(query_str);
-      if let Some(p_match) =
-        &self
-          .root_node()
-          .get_match_for_query(&self.code(), tree_sitter_scope_query, true)
-      {
-        return get_node_for_range(
-          self.root_node(),
-          p_match.range().start_byte,
-          p_match.range().end_byte,
-        );
-      }
-    }
-    self.root_node()
-  }
-
-  /// Apply all `rules` sequentially.
-  fn apply_rules(
-    &mut self, rules_store: &mut RuleStore, rules: &[Rule], parser: &mut Parser,
-    scope_query: Option<String>,
-  ) {
-    for rule in rules {
-      self.apply_rule(rule.to_owned(), rules_store, parser, &scope_query)
-    }
-  }
+  info!("Total files affected/matched {}", &summaries.len());
+  info!("Total number of matches {}", total_number_of_matches);
+  info!("Total number of rewrites {}", total_number_of_rewrites);
 }
 
 // Maintains the state of Piranha and the updated content of files in the source code.
