+++ polyglot/piranha/src/models/unit_tests/rule_test.rs
@@ -93,10 +93,10 @@ fn test_get_edit_positive_recursive() {
     rule_store.piranha_args(),
   );
   let node = source_code_unit.root_node();
-  let matches = rule.get_matches(&source_code_unit, &mut rule_store, node, true);
+  let matches = source_code_unit.get_matches(rule.clone(), &mut rule_store, node, true);
   assert!(!matches.is_empty());
 
-  let edit = rule.get_edit(&source_code_unit, &mut rule_store, node, true);
+  let edit = source_code_unit.get_edit(rule.clone(), &mut rule_store, node, true);
   assert!(edit.is_some());
 }
 
@@ -140,9 +140,9 @@ fn test_get_edit_negative_recursive() {
     rule_store.piranha_args(),
   );
   let node = source_code_unit.root_node();
-  let matches = rule.get_matches(&source_code_unit, &mut rule_store, node, true);
+  let matches = source_code_unit.get_matches(rule.clone(), &mut rule_store, node, true);
   assert!(matches.is_empty());
-  let edit = rule.get_edit(&source_code_unit, &mut rule_store, node, true);
+  let edit = source_code_unit.get_edit(rule.clone(), &mut rule_store, node, true);
   assert!(edit.is_none());
 }
 
@@ -179,13 +179,8 @@ fn test_get_edit_for_context_positive() {
     PathBuf::new().as_path(),
     rule_store.piranha_args(),
   );
-  let edit = Rule::get_edit_for_context(
-    &source_code_unit,
-    41_usize,
-    44_usize,
-    &mut rule_store,
-    &vec![rule],
-  );
+  let edit =
+    source_code_unit.get_edit_for_context(41_usize, 44_usize, &mut rule_store, &vec![rule]);
   // let edit = rule.get_edit(&source_code_unit, &mut rule_store, node, true);
   assert!(edit.is_some());
 }
@@ -224,13 +219,8 @@ fn test_get_edit_for_context_negative() {
     PathBuf::new().as_path(),
     rule_store.piranha_args(),
   );
-  let edit = Rule::get_edit_for_context(
-    &source_code_unit,
-    29_usize,
-    33_usize,
-    &mut rule_store,
-    &vec![rule],
-  );
+  let edit =
+    source_code_unit.get_edit_for_context(29_usize, 33_usize, &mut rule_store, &vec![rule]);
   // let edit = rule.get_edit(&source_code_unit, &mut rule_store, node, true);
   assert!(edit.is_none());
 }
